+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - ⛔ Disable `signin` command on macOS 10.13+ #162
 - 🐛 Fix `signout` command #162
+- 🐛 Fix `account` command
 - ➖ CocoaSeeds #155
 - ➕🍫 CocoaPods (1.5.3) #155
 - 🛠 Xcode 9.4 #153